requirements.txt
Django >= 2.2.1
psycopg2-binary >= 2.8.2
djangorestframework >= 3.9.4
djangorestframework-gis >= 0.14
phonenumbers >= 8.10.12
django-phonenumber-field >= 2.4.0
django-language-field >= 0.0.3
django-money >= 0.14.4